Stroy#1 When the user launches the application, he should land in a screen where is possible to see top headlines for the specific news source
This is completed.

How:
Used a Single Activity MVVM architecture. The NewsHeadlinesListFragment is responsible for showing the latest headlines from a specific source.
The data is provided by NewsHeadlinesViewModel.kt. It implements a paging object. This paging data is futher provided a lower level layer NewsHeadlinesPagingSource.kt which ultimately calls the API service.

This separation of concerns helps in better testing of things.