Story 2: When the user taps on a headline, he should be taken to a new screen
This is completed.

How:
The NewsDetailFragment is responsible for showing the details of selected news.
The data is provided by passing DomainModel via nav args . The data is set in viewmodel and then emitted to fragment by viewmodel via Kotlin Flow.

Test case for setting the item and emission is done and passed